+++ dataforge-io/src/commonTest/kotlin/space/kscience/dataforge/io/EnvelopeFormatTest.kt
@@ -25,6 +25,7 @@ class EnvelopeFormatTest {
         val bytes = res.data?.read {
             readByteArray()
         }
+//        println(byteArray.decodeToString())
         assertEquals("12345678", bytes?.decodeToString())
     }