+++ CHANGELOG.md
@@ -14,6 +14,20 @@
 
 ### Security
 
+## 0.10.2 - 2025-12-17
+
+### Added
+
+- Move Tables-kt to DataForge repository and make it follow DataForge versioning
+
+### Changed
+
+- **BREAKING** `Context.request(plugin)` now has nullable parameter null means taking existing plugin with any configuration.
+
+### Fixed
+
+- Some bugs in name parsing and validation.
+
 ## 0.10.1 - 2025-02-23
 
 ### Added